requirements.txt
# contextlib2
Cython
# lxml
matplotlib
numpy
opencv-python
pillow
tensorflow-gpu>=1.4, <2